Day24-24 part 2 ia done.
This part requires manual work from the user. Run it and use the
output file and the hint messages of the program to fix the broken wires.
Input the wires into the program and it will check if you are right.

When you are done, there should be only one false positive left, for the
first adder.